add aml sp
